Cythonized the grid class
Functions involved:
- _calculate_angular_distance
-  euclidean_distance
Note that the performence of cython is slower than the one of weave, depending of course, on the size of the grid. Note further, that both functions could also be vectorized. However the performence of the vectorized versions were still a bit slower the cython, but in the order of the cython version.

]]