Fix of double counting errors
The current implementation does not handle double countings correctly. This can quickly be reproduced for event series with events on alternating days. E.g.
EventSeriesX = [1,0,1,0,1,0,1,0,1,0,1,0,1,0,1,0]
EventSeriesY = [0,1,0,1,0,1,0,1,0,1,0,1,0,1,0,1]
produces an output of (0.83333333333333337, 1.0) which after the addition of both synchronisation rates is not normalised to (0,1) anymore. This is because events are counted twice due to the less or equal sign in both directions. To fix this, we have to check for all synchronised event pairs if at least one event of the pair is also synchronised in the other direction. For a given event pair, this can be done by looking at possible synchronisations in both columns and rows of the matrix of the other direction.  If any double countings are detected, the corresponding event pairs are only counted as 0.5 so that the normalisation works properly again.